Move `useSettings` to the hooks directory.
